refactor(ui): consolidate channel-subscription create/edit into one mode-prop form
