+++ services/dashboard-ui/client/components/slack/ChannelSubscriptionForm/index.ts
@@ -1,0 +1,1 @@
+export { ChannelSubscriptionFormModal } from './ChannelSubscriptionForm'
