+++ services/dashboard-ui/client/components/slack/CreateChannelSubscription/index.ts
@@ -1,2 +1,1 @@
 export { CreateChannelSubscriptionButton } from './CreateChannelSubscriptionContainer'
-export { CreateChannelSubscriptionModal } from './CreateChannelSubscription'
