+++ services/dashboard-ui/client/components/slack/EditChannelSubscription/index.ts
@@ -1,2 +1,1 @@
 export { EditChannelSubscriptionButton } from './EditChannelSubscriptionContainer'
-export { EditChannelSubscriptionModal } from './EditChannelSubscription'
